I-168: Change shared messages traits for objects
