+++ examples/src/main/scala/examples/hybrid/api/http/WalletApiRoute.scala
@@ -23,7 +23,7 @@ case class WalletApiRoute(override val settings: RESTApiSettings, nodeViewHolder
                          (implicit val context: ActorRefFactory)
   extends ApiRouteWithFullView[HybridHistory, HBoxStoredState, HWallet, SimpleBoxTransactionMemPool] {
 
-  import scorex.core.NodeViewHolder.ReceivableMessages.LocallyGeneratedTransaction
+  import scorex.core.LocallyGeneratedModifiersMessages.ReceivableMessages.LocallyGeneratedTransaction
 
   //TODO move to settings?
   val DefaultFee = 100
