+++ src/main/scala/scorex/core/LocallyGeneratedModifiersMessages.scala
@@ -3,7 +3,10 @@ package scorex.core
 import scorex.core.transaction.Transaction
 import scorex.core.transaction.box.proposition.Proposition
 
-trait LocallyGeneratedModifiersMessages {
-  case class LocallyGeneratedTransaction[P <: Proposition, TX <: Transaction[P]](tx: TX)
-  case class LocallyGeneratedModifier[PMOD <: PersistentNodeViewModifier](pmod: PMOD)
+// Messages shared by LocalInterface and NodeViewHolder
+object LocallyGeneratedModifiersMessages {
+  object ReceivableMessages {
+    case class LocallyGeneratedTransaction[P <: Proposition, TX <: Transaction[P]](tx: TX)
+    case class LocallyGeneratedModifier[PMOD <: PersistentNodeViewModifier](pmod: PMOD)
+  }
 }
