+++ src/main/scala/scorex/core/network/NetworkControllerSharedMessages.scala
@@ -3,6 +3,9 @@ package scorex.core.network
 import scorex.core.network.message.MessageSpec
 import scala.reflect.runtime.universe.TypeTag
 
-trait NetworkControllerSharedMessages {
-  case class DataFromPeer[DT: TypeTag](spec: MessageSpec[DT], data: DT, source: ConnectedPeer)
+// Messages shared by NetworkController, PeerSynchronizer and NodeViewSynchronizer
+object NetworkControllerSharedMessages {
+  object ReceivableMessages {
+    case class DataFromPeer[DT: TypeTag](spec: MessageSpec[DT], data: DT, source: ConnectedPeer)
+  }
 }
