+++ src/main/scala/scorex/core/network/PeerSynchronizer.scala
@@ -17,7 +17,7 @@ import scala.language.postfixOps
 
 class PeerSynchronizer(val networkControllerRef: ActorRef, peerManager: ActorRef, settings: NetworkSettings) extends Actor with ScorexLogging {
 
-  import PeerSynchronizer.ReceivableMessages._
+  import scorex.core.network.NetworkControllerSharedMessages.ReceivableMessages.DataFromPeer
   import scorex.core.network.peer.PeerManager.ReceivableMessages.{RandomPeers, AddOrUpdatePeer}
   import scorex.core.network.NetworkController.ReceivableMessages.{SendToNetwork, RegisterMessagesHandler}
 
@@ -54,10 +54,6 @@ class PeerSynchronizer(val networkControllerRef: ActorRef, peerManager: ActorRef
 
     case nonsense: Any => log.warn(s"PeerSynchronizer: got something strange $nonsense")
   }
-}
-
-object PeerSynchronizer {
-  object ReceivableMessages extends NetworkControllerSharedMessages {}
 }
 
 object PeerSynchronizerRef {
